refactor(mobile): split CelebrationModal into focused modules
Decomposes apps/mobile/src/components/ui/CelebrationModal.tsx (671 LOC)
into 9 focused modules under CelebrationModal/, mirroring the model
that landed in #1453 (TransactionsPage):

- CelebrationModal/CelebrationModal.tsx (orchestrator, 297 LOC)
- CelebrationModal/index.ts (re-exports, 16 LOC)
- CelebrationModal/types.ts (CelebrationType / ModuleTheme /
  ConfettiIntensity / ConfettiParticle / CelebrationModalProps /
  CelebrationConfig, 59 LOC)
- CelebrationModal/constants.ts (MODULE_COLORS / MODULE_BG_COLORS /
  CONFETTI_COUNTS, 31 LOC)
- CelebrationModal/haptics.ts (triggerHapticPattern, 61 LOC)
- CelebrationModal/confetti/createConfettiParticle.ts (factory, 32)
- CelebrationModal/confetti/ConfettiParticleView.tsx (particle view, 46)
- CelebrationModal/confetti/useConfetti.ts (hook, 72)
- CelebrationModal/hooks/useCelebration.tsx (convenience API, 138)

Public surface unchanged: CelebrationModal, useCelebration,
CelebrationModalProps, CelebrationType, ModuleTheme are re-exported
from index.ts so apps/mobile/src/components/ui/index.ts works as-is.

The unused _cx helper was dropped (dead code, no callers).

Verification:
- pnpm --filter @sergeant/mobile typecheck: 23 errors before / 23 after
  (all pre-existing in @sergeant/db-schema/migrate/sqlite resolution,
  none mention CelebrationModal).
- npx eslint src/components/ui/CelebrationModal/: clean (exit 0).
- npx vitest run src/components/ui: 8 pre-existing SyntaxError
  failures on Banner / Card / ConfirmDialog / Input / PageSkeleton /
  Sheet / Skeleton / Toast (transform pipeline issue on origin/main),
  none mention CelebrationModal. No CelebrationModal-specific tests
  exist in the repo (web canonical at apps/web also has no tests).

Behavioural parity: zero logic changes. Only re-grouped declarations
into smaller files and adjusted relative imports (../Button etc.)
